+++ frontend/app/api/export/lease/route.ts
@@ -1,6 +1,6 @@
 import { Role } from "@prisma/client";
 import { requireRole } from "../../../../services/auth";
-import { defaultLeaseSettings } from "../../../../util/leaseDefaults";
+import { defaultLeaseSettings } from "../../../../util/lease/leaseDefaults";
 import { formatDayColumn } from "../../../../util/recurrenceDisplay";
 import type { ILeaseSettings, IRoomRate } from "../../../../types/models";
 import { prisma } from "../../../../lib/prisma";
