+++ frontend/app/components/admin/export/ExportTab.tsx
@@ -13,7 +13,7 @@ import {
 import { ROOM_COLORS, ZOOM_ROOM_COLOR, CATEGORY_COLOR } from "../../../../util/filterColors";
 import type { ILeaseSettings } from "../../../../types/models";
 import { ALL_MEETING_EXPORT_FIELD_KEYS, type MeetingExportFieldKey } from "../../../../util/meetingExportFields";
-import type { LeaseYearCycle } from "../../../../util/leaseYearCycles";
+import type { LeaseYearCycle } from "../../../../util/lease/leaseYearCycles";
 import FilterGroup, { FilterGroupItem } from "../../shared/FilterGroup";
 import Card from "../shared/Card";
 import CardHeader from "../shared/CardHeader";
