+++ frontend/app/components/admin/export/LeaseConfigModal.tsx
@@ -4,7 +4,7 @@ import React, { useState } from "react";
 import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
 import CheckCircleIcon from "@mui/icons-material/CheckCircle";
 import type { ILeaseSettings, IRoomRate } from "../../../../types/models";
-import type { LeaseYearCycle } from "../../../../util/leaseYearCycles";
+import type { LeaseYearCycle } from "../../../../util/lease/leaseYearCycles";
 import styles from "../../../../styles/components/admin/ExportTab.module.scss";
 
 interface LeaseConfigModalProps {
